Use system call for munmap
